RNN's for univariate data
I explored the use of RNN's using Keras. I explored merging results of the different neural networks. In a sense it was like ensembling individual networks which might be a little overkill for the task but it allowed me to stabilize results since there was a lot of variability in results.  The weights never seemed to stabilize completely.